Add Quadricopter.ttm
First try to create a model with 2 more joints per arm. It should be modified later. In particular there should be only convex shape in the model for simulation reasons. Also it has a slight probem in the final orientation wrt the target object. The control part should be deleted after a control script in python will be developed. Mostly based on the current Quadricopter.ttm model in v-rep.